+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## FUTURE
 
+* Out-of-range internal_current_three values are no longer reported for AC300
+
 ## 0.13.0
 
 * Add initial logging support for EP600 (and very basic polling)