+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * BREAKING: The internal API has changed to better reflect standard MODBUS terminology
 * Out-of-range internal_current_three values are no longer reported for AC300
+* No longer checks if it can connect to all listed devices when logging
 
 ## 0.13.0